# Add regex processor plugin