--- a/plugins/processors/all/all.go
+++ b/plugins/processors/all/all.go
@@ -2,4 +2,5 @@ package all
 
 import (
 	_ "github.com/influxdata/telegraf/plugins/processors/printer"
+	_ "github.com/influxdata/telegraf/plugins/processors/regex"
 )