@44px 2018-04-08
Thanks for your interest!

As for now, this plugin operates only on values, so it can't convert tag to field or change its name.

I thought about some option which allows to convert between tag and field but later decided it would be better to have separate plugin for this. Such plugin can have more obvious name – "tag2field" instead of too general "regex" – so it would be easier to find it.

> extract tag name and value from partial field name (while renaming the field name too)

I'm not sure I properly understand your use case here. Can you give example of field name you have and what tag you want to create from it?