@voiprodrigo 2018-04-08
> extract tag name and value from partial field name (while renaming the field name too)

>> I'm not sure I properly understand your use case here. Can you give example of field name you have and what tag you want to create from it?

Actually, what I was thinking would require splitting the metric point into multiple ones, which should probably be handled by the plugin itself. Or be a stand-alone processor maybe.

Let's say a metric has field names which denote a thread index in a multi-threaded process: 

`measurement,pname=whatever thread0_valueA=0,thread1_valueA=0,thread0_valueB=1,thread1_valueB=2`

This would translate to:

´measurement,pname=whatever,thread=0 valueA=0,valueB=1
measurement,pname=whatever,thread=1 valueA=0,valueB=2`

So this would be something more or less like:
- Match all fields named ^(thread)(\d+)_(.+)$
- Create new metric points with new tag \1=\2
- Matched fields renamed to \3

I implement something like this as an option to an existing input plugin (targeting specific field names created by the plugin), so I was wondering if your plugin could be used as a generic way to the same kind of thing.