@danielnelson 2018-04-09
@voiprodrigo This seems like a fairly specialized operation, I'm not sure if it can be generalized nicely.  I think for this it would make sense if we had a processor that could run a user script to filter messages.  To avoid slowdown we would run the processor once and feed data in/out via stdin/stdout.