@44px 2018-05-20
Thanks for the tips! I applied `GetTag`/`GetField` and added cache for compiled regex patterns.

Here are results of running `go test -bench=. -benchmem` for this package:

Before:
```
pkg: github.com/influxdata/telegraf/plugins/processors/regex
BenchmarkConversions-4            100000             14945 ns/op           11192 B/op        147 allocs/op
PASS
ok      github.com/influxdata/telegraf/plugins/processors/regex 1.676s

```

After:
```
pkg: github.com/influxdata/telegraf/plugins/processors/regex
BenchmarkConversions-4            500000              2326 ns/op             632 B/op         25 allocs/op
PASS
ok      github.com/influxdata/telegraf/plugins/processors/regex 1.209s
```